add header and reformat
