+++ processing/src/main/java/io/druid/segment/data/BlockLayoutIndexedLongsSupplier.java
@@ -1,0 +1,185 @@
+/*
+ * Licensed to Metamarkets Group Inc. (Metamarkets) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Metamarkets licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.druid.segment.data;
+
+import com.google.common.base.Supplier;
+import com.google.common.io.Closeables;
+import com.metamx.common.guava.CloseQuietly;
+import io.druid.collections.ResourceHolder;
+
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.nio.LongBuffer;
+
+public class BlockLayoutIndexedLongsSupplier implements Supplier<IndexedLongs>
+{
+
+  private final GenericIndexed<ResourceHolder<ByteBuffer>> baseLongBuffers;
+  private final int totalSize;
+  private final int sizePer;
+  private final ByteOrder order;
+  private final CompressionFactory.LongEncodingFormatReader baseReader;
+
+  public BlockLayoutIndexedLongsSupplier(
+      int totalSize, int sizePer, ByteBuffer fromBuffer, ByteOrder order,
+      CompressionFactory.LongEncodingFormatReader reader,
+      CompressedObjectStrategy.CompressionStrategy strategy
+  )
+  {
+    baseLongBuffers = GenericIndexed.read(fromBuffer, VSizeCompressedObjectStrategy.getBufferForOrder(
+        order, strategy, reader.getNumBytes(sizePer)
+    ));
+    this.totalSize = totalSize;
+    this.sizePer = sizePer;
+    this.order = order;
+    this.baseReader = reader;
+  }
+
+  @Override
+  public IndexedLongs get()
+  {
+    final int div = Integer.numberOfTrailingZeros(sizePer);
+    final int rem = sizePer - 1;
+    final boolean powerOf2 = sizePer == (1 << div);
+    if (powerOf2) {
+      // this provide slightly better performance than calling the LongsEncodingReader.read, probably because Java
+      // doesn't inline the method call for some reason. This should be removed when test show that performance
+      // of using read method is same as directly accessing the longbuffer
+      if (baseReader instanceof LongsEncodingReader) {
+        return new BlockLayoutIndexedLongs()
+        {
+          @Override
+          public long get(int index)
+          {
+            // optimize division and remainder for powers of 2
+            final int bufferNum = index >> div;
+
+            if (bufferNum != currIndex) {
+              loadBuffer(bufferNum);
+            }
+
+            final int bufferIndex = index & rem;
+            return longBuffer.get(longBuffer.position() + bufferIndex);
+          }
+
+          protected void loadBuffer(int bufferNum)
+          {
+            CloseQuietly.close(holder);
+            holder = singleThreadedLongBuffers.get(bufferNum);
+            buffer = holder.get();
+            longBuffer = buffer.order(order).asLongBuffer();
+            currIndex = bufferNum;
+          }
+        };
+      } else {
+        return new BlockLayoutIndexedLongs()
+        {
+          @Override
+          public long get(int index)
+          {
+            // optimize division and remainder for powers of 2
+            final int bufferNum = index >> div;
+
+            if (bufferNum != currIndex) {
+              loadBuffer(bufferNum);
+            }
+
+            final int bufferIndex = index & rem;
+            return reader.read(bufferIndex);
+          }
+        };
+      }
+
+    } else {
+      return new BlockLayoutIndexedLongs();
+    }
+  }
+
+  private class BlockLayoutIndexedLongs implements IndexedLongs
+  {
+    final CompressionFactory.LongEncodingFormatReader reader = baseReader.duplicate();
+    final Indexed<ResourceHolder<ByteBuffer>> singleThreadedLongBuffers = baseLongBuffers.singleThreaded();
+    int currIndex = -1;
+    ResourceHolder<ByteBuffer> holder;
+    ByteBuffer buffer;
+    LongBuffer longBuffer;
+
+    @Override
+    public int size()
+    {
+      return totalSize;
+    }
+
+    @Override
+    public long get(int index)
+    {
+      final int bufferNum = index / sizePer;
+      final int bufferIndex = index % sizePer;
+
+      if (bufferNum != currIndex) {
+        loadBuffer(bufferNum);
+      }
+
+      return reader.read(bufferIndex);
+    }
+
+    @Override
+    public void fill(int index, long[] toFill)
+    {
+      if (totalSize - index < toFill.length) {
+        throw new IndexOutOfBoundsException(
+            String.format(
+                "Cannot fill array of size[%,d] at index[%,d].  Max size[%,d]", toFill.length, index, totalSize
+            )
+        );
+      }
+      for (int i = 0; i < toFill.length; i++) {
+        toFill[i] = get(index + i);
+      }
+    }
+
+    protected void loadBuffer(int bufferNum)
+    {
+      CloseQuietly.close(holder);
+      holder = singleThreadedLongBuffers.get(bufferNum);
+      buffer = holder.get();
+      currIndex = bufferNum;
+      reader.setBuffer(buffer);
+    }
+
+    @Override
+    public String toString()
+    {
+      return "BlockCompressedIndexedLongs_Anonymous{" +
+             "currIndex=" + currIndex +
+             ", sizePer=" + sizePer +
+             ", numChunks=" + singleThreadedLongBuffers.size() +
+             ", totalSize=" + totalSize +
+             '}';
+    }
+
+    @Override
+    public void close() throws IOException
+    {
+      Closeables.close(holder, false);
+    }
+  }
+}
